typify

Struct Type

pub struct Type<'a> { /* private fields */ }

Representation of a type which may have a definition or may be built-in.

Implementations

impl<'a> Type<'a>

pub fn name(&self) -> String

The name of the type as a String.

pub fn ident(&self) -> TokenStream

The identifier for the type as might be used for a function return or defining the type of a member of a struct..

pub fn parameter_ident(&self) -> TokenStream

The identifier for the type as might be used for a parameter in a function signature. In general: simple types are the same as Type::ident and complex types prepend a &.

pub fn parameter_ident_with_lifetime(&self, lifetime: &str) -> TokenStream

The identifier for the type as might be used for a parameter in a function signature along with a lifetime parameter. In general: simple types are the same as Type::ident and complex types prepend a &'<lifetime>.

pub fn describe(&self) -> String

A textual description of the type appropriate for debug output.

pub fn details(&self) -> TypeDetails<'_>

Get details about the type.

pub fn has_impl(&self, impl_name: TypeSpaceImpl) -> bool

Checks if the type has the associated impl.

pub fn builder(&self) -> Option<TokenStream>

Provides the the type identifier for the builder if one exists.