typify

Struct TypeSpace

pub struct TypeSpace { /* private fields */ }

A collection of types.

Implementations

impl TypeSpace

pub fn break_cycles(&mut self, range: Range<u64>)

We need to root out any containment cycles, breaking them by inserting a Box type. Our choice of where to break cycles is more arbitrary than optimal, but is well beyond sufficient.

impl TypeSpace

pub fn new(settings: &TypeSpaceSettings) -> TypeSpace

Create a new TypeSpace with custom settings

pub fn add_ref_types<I, S>(&mut self, type_defs: I) -> Result<(), Error>
where I: IntoIterator<Item = (S, Schema)>, S: AsRef<str>,

Add a collection of types that will be used as references. Regardless of how these types are defined–de novo or built-in–each type will appear in the final output as a struct, enum or newtype. This method may be called multiple times, but collections of references must be self-contained; in other words, a type in one invocation may not refer to a type in another invocation.

pub fn add_type(&mut self, schema: &Schema) -> Result<TypeId, Error>

Add a new type and return a type identifier that may be used in function signatures or embedded within other types.

pub fn add_type_with_name( &mut self, schema: &Schema, name_hint: Option<String>, ) -> Result<TypeId, Error>

Add a new type with a name hint and return a the components necessary to use the type for various components of a function signature.

pub fn add_root_schema( &mut self, schema: RootSchema, ) -> Result<Option<TypeId>, Error>

Add all the types contained within a RootSchema including any referenced types and the top-level type (if there is one and it has a title).

pub fn get_type(&self, type_id: &TypeId) -> Result<Type<'_>, Error>

Get a type given its ID.

pub fn uses_chrono(&self) -> bool

Whether the generated code needs chrono crate.

pub fn uses_regress(&self) -> bool

Whether the generated code needs [regress] crate.

pub fn uses_serde_json(&self) -> bool

Whether the generated code needs serde_json crate.

pub fn uses_uuid(&self) -> bool

Whether the generated code needs uuid crate.

pub fn iter_types(&self) -> impl Iterator<Item = Type<'_>>

Iterate over all types including those defined in this TypeSpace and those referred to by those types.

pub fn to_stream(&self) -> TokenStream

All code for processed types.